+++ src/test/resources/inspections/CollectFunctionInCollectionInspection/nonMatches/collect_on_no_type_variable.php
@@ -1,0 +1,5 @@
+<?php
+
+function($variable) {
+    collect($variable);
+}